# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 3.0.0
  description: Carter's Care v3 - NDIS & Aged Care Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /case-notes:
    get:
      operationId: getCaseNotes
      tags: [casenotes]
      summary: List case notes
      parameters:
        - name: participantId
          in: query
          schema:
            type: integer
        - name: staffId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of case notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseNote"
    post:
      operationId: createCaseNote
      tags: [casenotes]
      summary: Create case note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseNoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseNote"
  /case-notes/{id}:
    get:
      operationId: getCaseNote
      tags: [casenotes]
      summary: Get case note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case note detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseNote"
    patch:
      operationId: updateCaseNote
      tags: [casenotes]
      summary: Update case note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseNoteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseNote"
    delete:
      operationId: deleteCaseNote
      tags: [casenotes]
      summary: Delete case note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # INCIDENTS
...
  /goals:
    get:
      operationId: getGoals
      tags: [goals]
      summary: List participant goals
      parameters:
        - name: participantId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goal detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"

  # COMPLIANCE
  /compliance/checks:
    get:
      operationId: getComplianceChecks
      tags: [compliance]
      summary: List compliance check items
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of compliance checks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComplianceCheck"
    post:
      operationId: createComplianceCheck
      tags: [compliance]
      summary: Create compliance check
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceCheckInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceCheck"
  /compliance/checks/{id}:
    patch:
      operationId: updateComplianceCheck
      tags: [compliance]
      summary: Update compliance check
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceCheckUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceCheck"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, coordinator, support_worker, viewer]
        organisation:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
...
    CaseNote:
      type: object
      required: [id, participantId, staffId, category, content, createdAt]
      properties:
        id:
          type: integer
        participantId:
          type: integer
        participantName:
          type: ["string", "null"]
        staffId:
          type: integer
        staffName:
          type: ["string", "null"]
        shiftId:
          type: ["integer", "null"]
        category:
          type: string
          enum: [progress_note, goal_update, behaviour_support, health_update, personal_care, community_participation, communication, incident_follow_up, medication, daily_living]
        content:
          type: string
        goalId:
          type: ["integer", "null"]
        isConfidential:
          type: boolean
        requiresFollowUp:
          type: boolean
        followUpDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    CaseNoteInput:
      type: object
      required: [participantId, category, content]
      properties:
        participantId:
          type: integer
        shiftId:
          type: ["integer", "null"]
        category:
          type: string
          enum: [progress_note, goal_update, behaviour_support, health_update, personal_care, community_participation, communication, incident_follow_up, medication, daily_living]
        content:
          type: string
        goalId:
          type: ["integer", "null"]
        isConfidential:
          type: boolean
        requiresFollowUp:
          type: boolean
        followUpDate:
          type: ["string", "null"]
    CaseNoteUpdate:
      type: object
      properties:
        content:
          type: string
        category:
          type: string
          enum: [progress_note, goal_update, behaviour_support, health_update, personal_care, community_participation, communication, incident_follow_up, medication, daily_living]
        requiresFollowUp:
          type: boolean
        followUpDate:
          type: ["string", "null"]
...
    Goal:
      type: object
      required: [id, participantId, title, category, status, targetDate, createdAt]
      properties:
        id:
          type: integer
        participantId:
          type: integer
        participantName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [independence, community_participation, health_wellbeing, relationships, employment_education, communication, daily_living, safety, other]
        status:
          type: string
          enum: [active, achieved, on_hold, discontinued]
        progress:
          type: integer
        targetDate:
          type: string
        ndisAligned:
          type: boolean
        strategies:
          type: ["string", "null"]
        measurableOutcomes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    GoalInput:
      type: object
      required: [participantId, title, category, targetDate]
      properties:
        participantId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [independence, community_participation, health_wellbeing, relationships, employment_education, communication, daily_living, safety, other]
        targetDate:
          type: string
        ndisAligned:
          type: boolean
        strategies:
          type: ["string", "null"]
        measurableOutcomes:
          type: ["string", "null"]
    GoalUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, achieved, on_hold, discontinued]
        progress:
          type: integer
        strategies:
          type: ["string", "null"]
        measurableOutcomes:
          type: ["string", "null"]
    ComplianceCheck:
      type: object
      required: [id, title, category, framework, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [ndis_practice_standard, aged_care_quality_standard, worker_screening, training, policy, documentation, audit]
        framework:
          type: string
          enum: [ndis, aged_care, both]
        status:
          type: string
          enum: [compliant, non_compliant, under_review, not_applicable]
        dueDate:
          type: ["string", "null"]
        completedDate:
          type: ["string", "null"]
        assignedToId:
          type: ["integer", "null"]
        evidence:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ComplianceCheckInput:
      type: object
      required: [title, category, framework]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [ndis_practice_standard, aged_care_quality_standard, worker_screening, training, policy, documentation, audit]
        framework:
          type: string
          enum: [ndis, aged_care, both]
        dueDate:
          type: ["string", "null"]
        assignedToId:
          type: ["integer", "null"]
    ComplianceCheckUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [compliant, non_compliant, under_review, not_applicable]
        completedDate:
          type: ["string", "null"]
        evidence:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...