# lib/api-zod/src/generated/types/caseNoteUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Carter's Care v3 - NDIS & Aged Care Management Platform API
 * OpenAPI spec version: 3.0.0
 */
import type { CaseNoteUpdateCategory } from "./caseNoteUpdateCategory";

export interface CaseNoteUpdate {
  content?: string;
  category?: CaseNoteUpdateCategory;
  requiresFollowUp?: boolean;
  /** @nullable */
  followUpDate?: string | null;
}